ml-pipeline/requirements.txt
# CercaFungo ML Pipeline — Dependencies
# Install: pip install -r requirements.txt

# Detection & Training
ultralytics>=8.1.0
torch>=2.1.0
torchvision>=0.16.0

# Dataset management
roboflow>=1.1.0
albumentations>=1.3.0
Pillow>=10.0.0
opencv-python>=4.8.0

# Data & APIs
requests>=2.31.0
pyyaml>=6.0
pandas>=2.1.0
numpy>=1.24.0
flickrapi>=2.4.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
rich>=13.5.0
tqdm>=4.66.0

# Evaluation
scikit-learn>=1.3.0

# Export
onnx>=1.15.0
onnxruntime>=1.16.0
coremltools>=7.0

# Notebooks
jupyter>=1.0.0
ipywidgets>=8.1.0
